scripts/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argilla==1.5.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
boto3==1.26.102
botocore==1.29.102
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
dataclasses-json==0.5.7
Deprecated==1.2.13
dnspython==2.3.0
effdet==0.3.0
et-xmlfile==1.1.0
fastapi==0.95.0
filelock==3.10.7
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
geojson==2.5.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gptcache==0.1.11
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
huggingface-hub==0.13.3
humanfriendly==10.0
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.139
layoutparser==0.3.4
lit==16.0.0
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.2
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.4.0
pinecone-client==2.2.1
portalocker==2.7.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.14.0
pyowm==3.3.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.7.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
s3transfer==0.6.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
tiktoken==0.3.3
timm==0.6.13
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.7
unstructured-inference==0.3.2
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
Wand==0.6.11
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
